tests/submission/DavidCarlisle/math-parse02.html

<!DOCTYPE html>
<html>
<head>
<title>math in html: innerHTML</title>
<script src="/resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<h1>math in html: innerHTML</h1>

<!-- 
THIS TEST ASSUMES THAT
http://www.w3.org/Bugs/Public/show_bug.cgi?id=11204
IS RESOLVED
-->

<div id="log" style="display:block"></div>

<div style="display:none">
<div id="d1"><math><mi>x</mi></math></div>
</div>


<script>

test(function() {
assert_equals(document.getElementById("d1").firstChild.innerHTML,"<mi>x</mi>")
},"innerHTML defined on math.");


</script>

           

